zadatak2/requirements.txt
flask-sqlalchemy
sqlalchemy
flask-security
flask-wtf
gunicorn
flask_migrate
flask_script